+++ changelog.md
@@ -1,8 +1,8 @@
 # 0.3.2
 - Now train, dive, vote and statistic-based missions can only appear once
 - Deleted drinks from Eating mission, and created a new Drinking mission.
-- Added new tags, `#c:drinks` and `#brewinandchewin:drinks`
-- Added support for tags in missions.json
+- Added new tags, `#c:drinks`,`#brewinandchewin:drinks`, `#createdeco:catwalks`, `#createdeco:ladders`, `#create:ladders`, `#createdeco:hulls`, `#createdeco:bars`, `#create:bars` and `#create:scaffoldings`
+- - Added support for tags in missions.json
 
 # 0.2.7
 - Added compat with Create: Missions' addons